Hide user info (#22)
* Enable Service Provider Relocation (#17) (#4)

Schema Registry Client's basic HTTP Authentication support is
implemented through a ServiceProvider.  Without handling the fact that
relocating the schema client also relocates the service implementations,
no implementations are found when the client attempts to find a strategy
that matches an authentication source type specified through
basic.auth.credentials.source...

* Modify USER_INFO fields to use type Password instead of String

Kafka Connect logs connector configurations before launching
them, which is a problem if some of those configuration
properties happen to contain sensitive information that does
not belong in a log file, such as any Basic HTTP Authentication
credentials MirrorTool ahs been configured to make use of.

Kakfa Connect provides a `Password` data type that is always
masked on display.  It was relatifely simple to change both
the USER_INFO fields recently added to use PASSWORD instead
of STRING as their data types.

The URL field can sometimes also contain a password, when
the authentiation source is set to URL instead of USER_INFO.
There is no way to make these data types conditional, so
it is not possible to make URL of type PASSWORD if the
credential source is URL, while it is also of type STRING
if the credential source is not URL.  Since the credential
format when using URL is the same as it is when using
USER_INFO, and there is arguably a good reason to not mask
the rest of the URL, the URL fields continue to have type
String here.